[feat]: Allure reports in integration tests PoC
