+++ package.json
@@ -43,8 +43,8 @@
     "tsx": "^3.12.2",
     "type-fest": "^3.9.0",
     "typescript": "^5.2.2",
-    "vite": "^4.5.0",
-    "vitest": "^0.34.6",
+    "vite": "^5.1.5",
+    "vitest": "^1.3.1",
     "zx": "^7.2.0"
   }
 }
