+++ packages/client/test/integration/test-node/.gitignore
@@ -1,0 +1,1 @@
+allure-results
