+++ packages/client/test/integration/test-node/package.json
@@ -2,16 +2,18 @@
   "name": "@iroha2/client-test-node",
   "private": true,
   "scripts": {
-    "test": "cross-env DEBUG=@iroha2* vitest run main.spec"
+    "test": "cross-env DEBUG=@iroha2* vitest run",
+    "typecheck": "tsc --noEmit"
   },
   "devDependencies": {
     "@iroha2/client": "workspace:^7.0.0",
     "@iroha2/crypto-target-node": "workspace:^1.1.1",
     "@iroha2/test-configuration": "workspace:^",
     "@iroha2/test-peer": "workspace:^0.0.1",
+    "allure-vitest": "^2.14.0",
     "cross-env": "^7.0.3",
     "fp-ts": "^2.13.1",
     "immer": "^9.0.19",
-    "vite": "^4.5.0"
+    "vite": "^5.1.5"
   }
 }
