+++ packages/client/test/integration/test-node/test/setup.ts
@@ -1,0 +1,22 @@
+import { FREE_HEAP } from '@iroha2/crypto-core'
+import { setCrypto } from '@iroha2/client'
+import { crypto } from '@iroha2/crypto-target-node'
+import { Logger as ScaleLogger } from '@iroha2/data-model'
+import { afterAll, expect, beforeAll } from 'vitest'
+
+declare global {
+  // eslint-disable-next-line no-var,no-inner-declarations
+  var __clientTestNodeSetupOnce: boolean | undefined
+}
+
+// For some reason, `crypto` resets each time between test files, while Scale Logger sets once
+setCrypto(crypto)
+
+if (!globalThis.__clientTestNodeSetupOnce) {
+  new ScaleLogger().mount()
+  globalThis.__clientTestNodeSetupOnce = true
+}
+
+afterAll(() => {
+  expect(FREE_HEAP.size).toEqual(0)
+})
