--- packages/client/test/integration/test-node/vitest.config.ts
@@ -1,9 +1,0 @@
-import { defineConfig } from 'vitest/config'
-
-export default defineConfig({
-  test: {
-    watch: false,
-    reporters: 'basic',
-    threads: false,
-  },
-})
