+++ packages/client/test/integration/test-web/package.json
@@ -18,7 +18,7 @@
     "@iroha2/test-configuration": "workspace:^",
     "@iroha2/test-peer": "workspace:^0.0.1",
     "@scale-codec/util": "^1.1.2",
-    "@vitejs/plugin-vue": "^4.0.0",
+    "@vitejs/plugin-vue": "^5.0.4",
     "@vue-kakuyaku/core": "^0.4.3",
     "@vueuse/core": "^8.0.1",
     "cross-env": "^7.0.3",
@@ -27,7 +27,7 @@
     "fp-ts": "^2.13.1",
     "sass": "^1.52.1",
     "start-server-and-test": "^1.14.0",
-    "vite": "^4.5.0",
+    "vite": "^5.1.5",
     "vue": "^3.2.47"
   }
 }
