+++ packages/crypto/test/integration/test-web/package.json
@@ -13,6 +13,6 @@
     "@iroha2/crypto-target-web": "workspace:^1.0.0",
     "cypress": "^12.5.1",
     "start-server-and-test": "^1.14.0",
-    "vite": "^4.5.0"
+    "vite": "^5.1.5"
   }
 }
